refactor(frontend): resolve pr comments bb-303
