+++ apps/frontend/src/pages/quiz/libs/components/analyzing/libs/enums/text-animation-delay.enum.ts
@@ -1,6 +1,6 @@
 const TextAnimationDelay = {
 	BASE_DELAY: 0,
-	DELAY_MULTIPLIER: 0.1,
+	DELAY_MULTIPLIER: 0.08,
 } as const;
 
 export { TextAnimationDelay };
