+++ apps/frontend/src/pages/quiz/libs/components/analyzing/libs/helpers/helpers.ts
@@ -1,1 +1,1 @@
-export { createAnimatedSpans } from "./create-animated-spans/create-animated-spans.js";
+export { createAnimatedSpans } from "./create-animated-spans/create-animated-spans.helper.js";
