+++ apps/frontend/src/pages/quiz/libs/components/analyzing/libs/types/text.type.ts
@@ -1,0 +1,6 @@
+type TextElementType = {
+	firstParagraph: React.ReactNode[];
+	secondParagraph: React.ReactNode[];
+};
+
+export { type TextElementType };
