+++ apps/frontend/src/pages/quiz/libs/components/analyzing/libs/types/types.ts
@@ -1,0 +1,1 @@
+export { type TextElementType } from "./text.type.js";
